% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{fitted_prclmm}
\alias{fitted_prclmm}
\title{A fitted PRC LMM}
\format{
A list comprising step 2 and step 3 as obtained
during the estimation of the PRC LMM
}
\usage{
data(fitted_prclmm)
}
\description{
This list contains a fitted PRC LMM, where the CBOCP is
computed using 50 cluster bootstrap samples. It is
used to reduce the computing time in the example of
the function \code{performance_prclmm}
}
\examples{
data(fitted_prclmm)
ls(fitted_prclmm)
}
\references{
Signorelli, M., Spitali, P., Tsonaka, R. (in review). 
On the prediction of survival outcomes using longitudinal 
and high-dimensional (omic) data.
}
\seealso{
\code{\link{performance_prclmm}}
}
\author{
Mirko Signorelli
}
\keyword{datasets}
