% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_mlpmmfit.R
\name{summary.mlpmmfit}
\alias{summary.mlpmmfit}
\title{Extract model fits from step 1 of PRC-LMM}
\usage{
\method{summary}{mlpmmfit}(object, yname, ...)
}
\arguments{
\item{object}{the output of `fit_lmms`}

\item{yname}{a character giving the name of one of the longitudinal
outcomes modelled within one of the MLPMM}

\item{...}{additional arguments}
}
\value{
The model summary as returned by `summary.multlcmm`
}
\description{
Utility function to extract the MLPMM summaries from a model fit
obtained through `fit_mlpmms`
}
\references{
Signorelli, M. (2024). pencal: an R Package for the Dynamic 
Prediction of Survival with Many Longitudinal Predictors. 
To appear in: The R Journal. Preprint: arXiv:2309.15600

Signorelli, M., Spitali, P., Al-Khalili Szigyarto, C, 
The MARK-MD Consortium, Tsonaka, R. (2021). 
Penalized regression calibration: a method for the prediction 
of survival outcomes using complex longitudinal and 
high-dimensional data. Statistics in Medicine, 40 (27), 6178-6196.
DOI: 10.1002/sim.9178
}
\seealso{
\code{\link{fit_mlpmms}} and \code{\link[lcmm]{summary.multlcmm}}
}
\author{
Mirko Signorelli
}
